+++ modules/engine/src/main/scala/com/gsk/kg/engine/functions/FuncDates.scala
@@ -1,0 +1,3 @@
+package com.gsk.kg.engine.functions
+
+object FuncDates {}
