+++ modules/engine/src/main/scala/com/gsk/kg/engine/functions/FuncForms.scala
@@ -1,0 +1,100 @@
+package com.gsk.kg.engine.functions
+
+import org.apache.spark.sql.Column
+import org.apache.spark.sql.functions.not
+
+import com.gsk.kg.engine.functions.Literals.DateLiteral
+import com.gsk.kg.engine.functions.Literals.promoteNumericBoolean
+
+object FuncForms {
+
+  /** Performs logical binary operation '==' over two columns
+    * @param l
+    * @param r
+    * @return
+    */
+  def equals(l: Column, r: Column): Column = {
+    DateLiteral
+      .applyDateTimeLiteral(l, r)(_ === _)
+      .otherwise(
+        promoteNumericBoolean(l, r)(_ === _)
+          .otherwise(l === r)
+      )
+  }
+
+  /** Peforms logical binary operation '>' over two columns
+    * @param l
+    * @param r
+    * @return
+    */
+  def gt(l: Column, r: Column): Column =
+    DateLiteral
+      .applyDateTimeLiteral(l, r)(_ > _)
+      .otherwise(
+        promoteNumericBoolean(l, r)(_ > _)
+          .otherwise(l > r)
+      )
+
+  /** Performs logical binary operation '<' over two columns
+    * @param l
+    * @param r
+    * @return
+    */
+  def lt(l: Column, r: Column): Column =
+    DateLiteral
+      .applyDateTimeLiteral(l, r)(_ < _)
+      .otherwise(
+        promoteNumericBoolean(l, r)(_ < _)
+          .otherwise(l < r)
+      )
+
+  /** Performs logical binary operation '<=' over two columns
+    * @param l
+    * @param r
+    * @return
+    */
+  def gte(l: Column, r: Column): Column =
+    DateLiteral
+      .applyDateTimeLiteral(l, r)(_ >= _)
+      .otherwise(
+        promoteNumericBoolean(l, r)(_ >= _)
+          .otherwise(l >= r)
+      )
+
+  /** Performs logical binary operation '>=' over two columns
+    * @param l
+    * @param r
+    * @return
+    */
+  def lte(l: Column, r: Column): Column =
+    DateLiteral
+      .applyDateTimeLiteral(l, r)(_ <= _)
+      .otherwise(
+        promoteNumericBoolean(l, r)(_ <= _)
+          .otherwise(l <= r)
+      )
+
+  /** Performs logical binary operation 'or' over two columns
+    * @param l
+    * @param r
+    * @return
+    */
+  def or(l: Column, r: Column): Column =
+    l || r
+
+  /** Performs logical binary operation 'and' over two columns
+    * @param r
+    * @param l
+    * @return
+    */
+  def and(l: Column, r: Column): Column =
+    l && r
+
+  /** Negates all rows of a column
+    * @param s
+    * @return
+    */
+  def negate(s: Column): Column =
+    not(s)
+
+}
